update all torch bincounts
